# backend/docs/swagger.yaml
basePath: /api
definitions:
  apiserver.CalcEstimate:
    properties:
      estimate:
        $ref: '#/definitions/apiserver.Estimate'
        format: Estimate
      hint:
        example: not all users provided estimates
        format: string
        type: string
      message:
        example: warning
        format: string
        type: string
      users:
        example:
        - Tigger
        format: '[]string'
        items:
          type: string
        type: array
    type: object
...
  apiserver.ErrorResponse:
    properties:
      message:
        example: error
        format: string
        type: string
      reason:
        example: oops, something went wrong
        format: string
        type: string
    type: object
  apiserver.Estimate:
    properties:
      effort:
        example: 1.5
        format: float64
        type: number
      standarddeviation:
        example: 0.2
        format: float64
        type: number
    type: object
  apiserver.GeneralResponse:
    properties:
      message:
        example: ok
        format: string
        type: string
      route:
        example: /sessions/token
        format: string
        type: string
    type: object
  apiserver.PerUserEstimate:
    properties:
      b:
        example: 1.5
        format: float64
        type: number
      id:
        example: TEST01
        format: string
        type: string
      m:
        example: 2
        format: float64
        type: number
      user:
        example: Tigger
        format: string
        type: string
      w:
        example: 3.6
        format: float64
        type: number
    type: object
  apiserver.PerUserEstimateResponse:
    properties:
      estimates:
        format: '[]datastore.Estimate'
        items:
          $ref: '#/definitions/datastore.Estimate'
        type: array
      message:
        example: ok
        format: string
        type: string
    type: object
  apiserver.User:
    properties:
      name:
        example: Tigger
        format: string
        type: string
    type: object
  apiserver.UsersResponse:
    properties:
      message:
        example: ok
        format: string
        type: string
      users:
        example:
        - Tigger
        - Rabbit
        format: '[]string'
        items:
          type: string
        type: array
    type: object
  apiserver.WorkPackage:
    properties:
      id:
        example: TEST01
        format: string
        type: string
      summary:
        example: a sample task
        format: string
        type: string
    type: object
  apiserver.WorkPackagesResponse:
    properties:
      message:
        example: ok
        format: string
        type: string
      workpackages:
        format: '[]datastore.WorkPackage'
        items:
          $ref: '#/definitions/datastore.WorkPackage'
        type: array
    type: object
  datastore.Estimate:
    properties:
      bestCase:
        type: number
      mostLikelyCase:
        type: number
      userName:
        type: string
      workPackageID:
        type: string
      worstCase:
        type: number
    type: object
  datastore.WorkPackage:
    properties:
      effort:
        type: number
      id:
        type: string
      standardDeviation:
        type: number
      summary:
        type: string
    type: object
host: localhost:5000
info:
  contact:
    name: HaRo87
    url: https://github.com/HaRo87
  description: A backend for playing Planning Poker with Delphi estimate method.
  license:
    name: MIT
    url: https://github.com/HaRo87/dokerb/blob/main/LICENSE
  title: Doker Backend API
  version: 0.1.0
paths:
...
  /sessions:
    post:
      description: Creates a new Doker session and responds with the corresponding
        token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Create a new Doker session
      tags:
      - session
  /sessions/{token}:
    delete:
      description: Deletes a existing Doker session based on the provided token
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Delete a existing Doker session
      tags:
      - session
  /sessions/{token}/estimates:
    get:
      description: Gets all estimates of all existing users of all existing work packages
        inside a existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.PerUserEstimateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Get the estimates of all users for all work packages
      tags:
      - estimate
    post:
      description: Adds a estimate of a existing user of a existing work package inside
        a existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: New Estimate
        in: body
        name: estimate
        required: true
        schema:
          $ref: '#/definitions/apiserver.PerUserEstimate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Add the estimate of a user for a work package
      tags:
      - estimate
  /sessions/{token}/estimates/{id}:
    get:
      description: Gets the average estimate of all existing users of a existing work
        package inside a existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: Work Package ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.CalcEstimate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Get the average estimate of all users for a specific work package
      tags:
      - estimate
  /sessions/{token}/estimates/{user}/{id}:
    delete:
      description: Removes a estimate of a existing user of a existing work package
        inside a existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: User Name
        in: path
        name: user
        required: true
        type: string
      - description: Work Package ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Remove the estimate of a user for a work package
      tags:
      - estimate
  /sessions/{token}/users:
    get:
      description: Gets all users of an existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.UsersResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Get the users of an existing session
      tags:
      - user
    post:
      description: Adds a new (non-existing) user to an existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: New User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apiserver.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Add a new user to a existing session
      tags:
      - user
  /sessions/{token}/users/{name}:
    delete:
      description: Removes a existing user from an existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: Name of the user
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Remove a user from a session
      tags:
      - user
  /sessions/{token}/workpackages:
    get:
      description: Gets all work packages of an existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.WorkPackagesResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Get the work packages of a session
      tags:
      - workpackage
    post:
      description: Adds a new (non-existing) work package to an existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: New Work Package
        in: body
        name: workpackage
        required: true
        schema:
          $ref: '#/definitions/apiserver.WorkPackage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Add a new work package to a existing session
      tags:
      - workpackage
  /sessions/{token}/workpackages/{id}:
    delete:
      description: Removes a existing work package from an existing session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: ID of the work package
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Remove a work package from a session
      tags:
      - workpackage
    put:
      description: Updates a estimate of a existing work package inside a existing
        session
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: ID of the work package
        in: path
        name: id
        required: true
        type: string
      - description: New Estimate
        in: body
        name: estimate
        required: true
        schema:
          $ref: '#/definitions/apiserver.Estimate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Update the estimate of a work package
      tags:
      - workpackage
  /sessions/{token}/workpackages/{id}/estimate:
    delete:
      description: Removes the estimate from an existing work package
      parameters:
      - description: Session Token
        in: path
        name: token
        required: true
        type: string
      - description: ID of the work package
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apiserver.GeneralResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apiserver.ErrorResponse'
      summary: Delete the estimate from a work package
      tags:
      - workpackage
swagger: "2.0"